% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{bababos_theme}
\alias{bababos_theme}
\title{This is official theme for bababos publication.}
\usage{
bababos_theme(base_theme = theme_gray, base_size = 8)
}
\arguments{
\item{base_theme}{Theme used as a base for the ggplot theme}

\item{base_size}{The size of the various text}
}
\value{
ggplot2 theme
}
\description{
This is official theme for bababos publication.
}
\author{
Suberlin Sinaga
}
