% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{create_pallete}
\alias{create_pallete}
\title{create_pallete}
\usage{
create_pallete(pallete)
}
\arguments{
\item{pallete}{Named vector of pallete to use}
}
\value{
Vector with pallete options
}
\description{
A pallete to create bababos brand and visualization scheme
}
\author{
Suberlin Sinaga
}
