% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{price_range_vis}
\alias{price_range_vis}
\title{price_range_vis}
\usage{
price_range_vis(
  df_summarized,
  commodity,
  importer = "importer",
  min_price = "min_price",
  max_price = "max_price"
)
}
\arguments{
\item{df_summarized}{data.frame: Summarized dataframe to importer level, that has min_price and max_price varilable}

\item{commodity}{str: The name of the entity that is being plotted}

\item{importer}{str: The column name stored the importer/exporter name}

\item{min_price}{str: The column name stored the minimum price}

\item{max_price}{str: The column name stored the minimum price}
}
\description{
price_range_vis
}
